Update dll_main.cpp
adds simple flag checking if we are named reshade32/reshade64, and that we are not using the VR/vulkan installation in programdata. In that case we can safely assume we are not using reshade as a wrapper so we want to support injection from anywhere. Currently in this case we'll still get the game dir as our base path which can cause a number of issues